% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{Data}
\alias{Data}
\title{Retrieve Data Object}
\usage{
Data(d, which = c("wd", "user", "system"), verbose = TRUE, envir = .GlobalEnv)
}
\arguments{
\item{d}{A dataset to load. Must be in the form \code{pkg::data}.}

\item{which}{One of c("wd","user","system"): current working directory (wd),
user config directory (user) as determined by \code{\link[rappdirs]{user_config_dir}},
or system package directory (system) as determined by \code{\link[base]{system.file}}.}

\item{verbose}{logical. Should extra information be printed during execution.}

\item{envir}{environment. The environment to load data into (default is .GlobalEnv).}
}
\value{
A character vector of the names of objects created, invisibly.
}
\description{
Analagous to \code{\link[utils]{data}}, but supporting indirection when loading via
configuration files.
}
\details{
The goal of this function is to provide a replacement for the \code{\link[utils]{data}}
command in R. Or more accurately, to wrap the command in a layer of indirection such
that even when using the command it is possible to retrieve a remote dataset.

The current strategy for retrieving data is to use the format \code{pkg::data}. This function
will search for a \code{conf} subdirectory of an installed package (\code{pkg}) and retrieve the
file \code{config.yml} using the \code{\link{config}{get}} method. This config file should
contain an entry \code{d}, which matches the argument of this function. The entry should
point to a URI suitable for executing the \code{\link[utils]{data}} function.
}
\examples{
\dontrun{
Data("pgreportr::grants")
}
}
